Update agent service management defaults
Before this, all platforms that were not Darwin, and this resulted in
the default case, would fail if there was no defined parameter for the
agent_service_conf variable.  It looks like what was attempted was to
remove the file if the variable was undef, but in that case, and undef
variable would result in an undef file path, thus not accomplishing the
goal.

This code alters the behaviour of the default case for agent service
management, so that it will at least succeed when the variable is not
present, and manage properly when the variable is.  This should be
sufficient for both cases.